+++ apps/service/src/components/problem/index.ts
@@ -1,0 +1,3 @@
+import ProblemCalandar from './ProblemCalandar';
+
+export { ProblemCalandar };
